pyproject.toml
[project]
name = "ta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "markdown>=3.8",
    "markdownify>=1.1.0",
    "numpy>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]
